+++ Movie Evening/CustomView/SectionHeaderView.swift
@@ -26,5 +26,6 @@ struct SectionHeaderView: View {
             }
         }
         .padding(.horizontal)
+        
     }
 }
