--- Movie Evening/Resources/CustomView/CollectionView.swift
@@ -1,43 +1,0 @@
-//
-//  CardMovie.swift
-//  Movie Evening
-//
-//  Created by Иван Тарасенко on 21.07.2023.
-//
-
-import SwiftUI
-
-//struct CardMovie: View {
-//    let title: String
-//    var body: some View {
-//        ZStack {
-//            VStack {
-//                RoundedRectangle(cornerRadius: 10)
-//                    .foregroundColor(.random)
-//                    .frame(height: 270)
-//                
-//            }
-//            
-//            Text(title)
-//                .font(.title2)
-//            Text("test")
-//        }
-//        
-//    }
-//}
-//
-//extension Color {
-//    static var random: Color {
-//        return Color(
-//            red: .random(in: 0...1),
-//            green: .random(in: 0...1),
-//            blue: .random(in: 0...1)
-//        )
-//    }
-//}
-
-
-
-//#Preview {
-//    CollectionView(title: "")
-//}
