+++ Movie Evening/Screens/Movies/Views/DetailAboutFilm.swift
@@ -9,7 +9,7 @@ import SwiftUI
 
 struct DetailAboutFilm: View {
     var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        Image(._360)
     }
 }
